moved hook for verify status repo access into hooks file; updated readme instructions regarding authenticated status requests; made minor syntax and style updates to various source files
