+++ types/http-signature/index.d.ts
@@ -1,2 +1,2 @@
-declare module "http-signature";
-declare module "http-signature/lib/utils";
+declare module 'http-signature';
+declare module 'http-signature/lib/utils';
